Machine_Learning/King_County_houses_prices/requirements.txt
flask
scikit-learn
pandas
numpy
joblib
mysql-connector-python
python-dotenv
